Low: libcrmcommon: Parse iso8601 strings according to spec.
The specification states that there should not be a space between the
time and the offset.  At the moment, however, we only parse the string
correctly if there is a space.  Since we've been doing it that way
forever, we can't get rid of that.  The best we can do is support both
the right way and the old way, and eventually get rid of the old way.