Refactor: tools: Call out->err() in local_files_set.
This prevents us from having to figure out what the error was in the
caller where we have less information.  This should also be the last of
the mid-level functions that needs to call out->err() since for the most
part, their errors occur when calling run_command or some other lower
level function.