Low: tools: cibsecret should work with CIB_file set.
All that's needed to make this work is to rearrange setting up the CIB
connection a little bit so we can then check if we have a cib_file
variant and if so, don't call pcmk__query_node_name.

This will mean we have no local node name (it's set to NULL), so the
list of nodes pulled from the CIB won't be filtered at all.

I only tested the set command, but I see no reason why other commands
shouldn't work.  The commands that involve syncing files between cluster
nodes will require the nodes to be running and for you to have
passwordless ssh set up for accessing those nodes.  This may or may not
make sense depending on what nodes are defined in the CIB_file.

I ran this locally by setting PATH=tools/ CIB_file=cts/cli/crm_mon.xml
and then running the command as root.

Fixes T993