API: libcrmcommon: Add pcmk_rc_cs_internal error code.
Unfortunately, a lot of corosync functions can return CS_ERR_LIBRARY for
a variety of error conditions so we need this more generic error code to
map to.  At the same time, we may want to differentiate corosync errors
from other kinds of errors, hence why this is separate from
pcmk_rc_error.

Ref T907