Refactor: libcrmservice: Rename services_os_get_single_directory_list()
The name is overly long and uses the public API prefix ("services_" with
a single underscore). Also rename the lpc variable to "i" and scope it
to the loop, rename the "root" argument to "dir", and use bool instead
of gboolean.

I'm guessing that the "os" prefix, as well as the separate
"services_linux.c" file, were meant to allow different implementations
of certain functions depending on the OS. GLib does similar things.
However, the commit that introduced it all (cc2110d7) sheds no light on
this question.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>