Low: build, libcrmservice: initdir must be a single directory
Previously, the libcrmservice code supported initdir being a
colon-delimited list of directories. However, this was not documented in
the configure help text or comments. All of the documentation referred
to initdir being a single directory.

The default behavior, if not building with "--with-initdir=<dir>", was
to check a list of directories and set INITDIR to the first one that
existed.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>