Refactor: fencer: Use g_strsplit() for pcmk_delay_base mappings
It's still confusing, but hopefully a little bit less so.

When delay_base_s is set, it means we've found our mapping. So we use
that as a loop exit condition.

We no longer need valptr, since we never change what value points to.

Now that value isn't changing, it will be easier to feel confident about
using g_strsplit_set() in place of strtok().

Note the similarity to pcmk__scan_nvpair().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>