Refactor: fencer: Use g_strsplit_set() in get_action_delay_base()
Instead of strtok(). This seems clearer.

Skip empty mappings so that we don't log a "malformed" error. strtok()
skipped contiguous sequences of delimiters, so we wouldn't try to parse
an empty mapping. g_strsplit_set() returns an empty string between
contiguous delimiters.

Strip leading and trailing whitespace before parsing the param value.

It seems reasonable not to strip leading or trailing semicolons. We can
ignore leading or trailing whitespace, but a leading or trailing
semicolon seems like a clear indicator that this is supposed to be a
mapping. This will result an error for values like ";3s " or " 3s;" but
not for " 3s ".

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>