Low: fencer: Fix ISO 8601 interval parsing in pcmk_delay_base
An ISO 8601 interval can include a colon character. I can't think of any
valid reason why a user would specify an interval this way for a fencing
delay, but there doesn't seem to be any reason to exclude mapping values
that have a colon. The "make sure there's no colon" check seemed to be
an artifact of the way the parsing code worked before we started using
g_strsplit(), etc.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>