Low: fencer: Drop support for escaped characters in pcmk_host_map
This reverts e95198f.

Backslash escapes are useful in C and on the command line, but it's not
clear how they would be useful in a pcmk_host_map value. Backslash
escapes are not used in XML, where the configuration is stored.

The removed code did the following upon finding a backslash:
* If we're not inside the value part of a host-to-port mapping, skip
  both a backslash and the character after it.
* Otherwise, skip the backslash and keep the character after it (unless
  that character is also a backslash).

This doesn't seem to make sense. Any characters that are special in XML
attribute values would need to be escaped using XML entities. Other
characters, including backslashes, are allowed as literals.
* https://www.w3.org/TR/REC-xml/#NT-AttValue

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>