Feature: libcrmcommon: Deprecate enable-startup-probes cluster property
Setting this property to false is dangerous in that it prevents
Pacemaker from gaining an accurate view of resource state, without
preventing other actions as (for example) maintenance mode would do.

Disable probes using a location constraint rule with
resource-discovery=never instead.

This option was introduced by commit b20fd762, and the expressed
motivation was that the "calculation [of startup probes] is a major
bottleneck for very large clusters." I doubt that's a major concern on
modern hardware, and even if it is, Pacemaker still should have a
correct view of resource state.

Ref T1000

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>