Low: daemons: Fix removing transient attrs on rolling upgrades.
The previous patches to fix race conditions in removing transient
attributes rely on this sequence of events:

* First, attrd notices that a peer has left the cluster at the corosync
  level.  This causes it to set the values of all transient attributes
  to NULL.
* Second, based updates the CIB to remove transient attributes, which
  triggers a callback in attrd.  This callback removes all attributes
  whose value is NULL.

This seems to work reliably in same-version clusters.  However, in mixed
version clusters (like if you're doing a rolling upgrade), this doesn't
always happen.  In fact in my testing, almost every time, the order of
those two operations is reversed.

What's happening here is that in the mixed version cluster, the DC
(which will still be the old, un-upgraded cluster version) still has a
controld that deletes transient attributes.  It will be sending out
requests that based performs this change to the CIB, which conflicts
with the entire point of these patches - that controld is no longer in
the business of deleting transient attributes.

I think there's no way around this problem.  No matter how you update
cluster nodes, you'll always have a DC that's the old version with an
old controld.  If you take the existing DC out of the cluster to update
it, another existing node will become the DC and it will be the problem
node.

The fix here is to bump the attrd protocol version so we know if a node
has been updated or not.  This is probably reasonable to do for the
change in the first place.  Then, attrd can inspect the protocol version
of the peer that is telling it to remove a transient attribute.  If the
peer version is old enough, attrd can simply remove the transient
attribute immediately instead of doing the NULL-then-delete process.