API: libcib: Meaningful output from XPath query for an attribute
With 5fe98f4d25 in 3.0.1, a matching attribute was displayed with a vague
element name "xpath-query". Besides, if there were multiple matches for an
attribute, only the last one would be displayed.

Now a matching attribute is displayed with the corresponding element name
and the ID if present:

$ cibadmin --query --xpath "/cib/@crm_feature_set"
<cib crm_feature_set="3.20.5"/>

Multiple matches are all displayed:

$ cibadmin --query --xpath "//@uname"
<xpath-query>
  <node id="1" uname="node1"/>
  <node id="2" uname="node2"/>
  <node_state id="1" uname="node1"/>
  <node_state id="2" uname="node2"/>
</xpath-query>