Refactor: controller: Drop fsa_data_t:data_type member
If data is not NULL, it's always set to fsa_dt_ha_msg. If data is NULL,
then it may be set to either fsa_dt_none or fsa_dt_ha_msg... probably
always fsa_dt_none, but this is not clearly the case.

In any case, we only need to check the data member.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>