Refactor: controller: Drop A_LRM_INVOKE
Nothing set it as a global action flag or used it in a meaningful way.
It was only passed to do_lrm_invoke(), which ignored it.

This also means that s_crmd_fsa_actions() can no longer call
do_lrm_invoke(). Rename it to controld_invoke_execd() and drop the
unused arguments.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>