Fix: libcrmcommon: Don't assert in time_to_hr.
This just allows time_to_hr and pcmk__time_hr_now to return NULL, which
can then be handled higher up in the call stack.  This is just to
address an automatically filed fuzzer bug report - it can't figure out
how to deal with the fact that a function it's fuzzing can assert.

Fixes https://issues.oss-fuzz.com/issues/456526118