Refactor: libcrmcommon: Drop pcmk__time_hr_new()
The only callers were a fuzzer line (which was only there to test this
function) and a unit test of a different function. We can just do the
assignment inside the unit test.

This leaves a little bit of duplication with time_to_hr(), but the goal
is to replace most of our iso8601 library with something external (see
upstream task T432). So I'm not concerned about duplication in one
place.

This does requires exposing struct crm_time_s internally, so that we can
reference its members in the unit test. Alternatively we could create
getter functions for the members, but this is easier.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>