Refactor: libcrmcommon: pcmk__time_format_hr() takes crm_time_t and int
A pcmk__time_hr_t object is just a crm_time_t with an int appended for
useconds. Instead of taking that object, having to construct it, etc.,
just take the crm_time_t and int.

This removes the sole use of pcmk__time_hr_t, so that we can drop it.
This makes sense to do, given that our goal is to replace most of our
time library with something external (see T432).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>