Refactor: controller: Explicit msg_data arg in controld_fsa_stall()
I find it easier to understand when it relies less on implicit context.
Prior to this commit, it required that a "msg_data" variable be defined
in the scope of the call, and the macro then cast it to (fsa_data_t *).

Now, we have the caller pass msg_data as an argument if appropriate; the
macro now calls it cur_data (similar to register_fsa_error()).

We also drop the suppress argument as redundant. Passing NULL for
cur_data produces the same behavior as suppress=TRUE previously did.
Passing msg_data for cur_data produces the same behavior as
suppress=FALSE previously did.

NOTE: It is not clear to us what the purpose of the suppress argument
was, and thus whether it's correct to pass NULL for cur_data now. This
commit simply preserves existing behavior.

The existing behavior was added by commit a9a81618 in 2013. That commit
adds the suppress argument. Via that argument, it starts implicitly
passing in msg_data at most call sites -- the ones with suppress=FALSE.
At certain other call sites -- all the ones in do_started(), and no
others -- the commit passes suppress=TRUE, so that msg_data is treated
as NULL.

The commit message:
* Is misleading. It says "Always store the data when stalling the FSA."
  It should say "sometimes." It does not store the data when stalling
  the FSA in do_started(). The decision of whether to store the data is
  made at the caller level as of that commit. Prior to that commit, the
  data was *never* stored when stalling the FSA.
* Gives no explanation of *why* it suppresses storing the data within
  do_started(). So we don't know whether that's appropriate or whether
  we should preserve that behavior.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>