Refactor: daemons: Remove a logging block from cib_process_request.
I'm not sure that we care about any of these logging messages.  I can't
remember ever looking for them in a log file.  I'm also not sure how
useful they are.  "Directing reply to..." is already covered by a
message in send_peer_reply.  The "Request not broadcast..." messages are
a little confusing when send_peer_reply is going to be called anyway.

That just leaves "Completed update as secondary", which I am sure I've
definitely never cared about.