Refactor: daemons: Simplify the do_local_notify call.
This isn't strictly necessary (and I kind of hate the ternary operator
in C in the first place), but there's just entirely too many conditional
blocks in cib_process_request.  I'm hoping this makes it a little less
intimidating to look at, at the cost of having to remember the ternary
operator is there.