Refactor: daemons: Split up IPC code in schedulerd.
The idea behind this commit (and all the ones that will follow it) is to
make all the daemons look as similar as possible when it comes to the
IPC code.  This will make it easier in the future to find code that's
common across daemons and condense into a single set of code that they
can all share.  I think there's probably quite a bit of this code, but
it's a little difficult to see at the moment.

This commit is a little pointless given how simple the IPC in the
scheduler is, but that's just how this is going to go.  The idea is that
daemon_ipc.c holds the code specific to IPC messages (as opposed to CPG
messages), and daemon_messages.c will hold the code common to both types
of messages.

There's still quite a bit wrong with the code organization right now,
but I'm going to fix that up over several commits.