Refactor: daemons: Standardize handling IPC requests in schedulerd.
This is more refactoring to bring this code in line with how the same
things are happening in attrd in order to find common code.

* schedulerd_ipc_dispatch shouldn't register message handlers, call
  pcmk__process_request, or deal with populating the result object.  All
  that should be done in a new schedulerd_handle_request function.

* schedulerd_handlers and the registration function can now be static
  since they're not needed elsewhere.

If this seems a little overdesigned or pointless here, remember that
other daemons support both IPC and CPG messages (schedulerd only does
IPC messages) and this is how they are organized.