Refactor: libcrmcommon: Process PCMK__XA_HIDDEN more robustly
This feature was introduced by efbede9 and is undocumented. Currently
the only thing that uses it is the CIB remote client, which sets it to
PCMK__XA_PASSWORD.

So we could actually do a direct string match. But for extensibility,
I'm treating the value as a comma-separated list of hidden attributes,
so that a client can pass multiple attribute names to be hidden.

It may be harder to add this feature if we do it later, because older
servers won't recognize multi-name values from newer clients.

In any case, using strstr() matched too broadly.

Also, go ahead and call pcmk__xml_escape() even on the constant "*****"
that we use for hidden values. It simplifies the code.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>