Refactor: daemons: Split out checking for a previous instance.
Most daemons have some way of doing this, but for the most part they are
pretty simple and therefore will be easy to combine.  The pacemakerd
version is special and does a lot of extra stuff, which makes it seem
unrealistic to merge into a single generic version.  However, splitting
it out into its own function at least makes the main function simpler
for when it comes time to standardize those.