Refactor: daemons: Clean up header files in schedulerd.
Going back and redoing the header files one commit at a time is more
work than it's worth, as far as I'm concerned.  Thus the previous
commits will just be left alone with their incorrect (but compilable)
includes.

This started out as just correcting the headers in the files I touched
as part of the IPC cleanup, but schedulerd is small enough that I might
as well correct the other files while I'm in the area.