Refactor: controller: Drop unescape_newlines()
We escape newlines as "\\n" (or "\\\\n" in C) only if
pcmk__xml_escape_attr_pretty is set. In the past, when
unescape_newlines() was added via c8206514, we escaped all newlines that
way in XML attributes.

It's not immediately clear what would have caused newlines in action
stdout and stderr to get escaped back then, but the important thing is
that nothing appears to do it now.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>