Refactor: include: Move function definitions into agents_internal.h.
This is the first part of a project to reduce crm/common/internal.h to a
file that basically just includes all the other internal files.  After
that, we can then use the #error technique to make sure the subheaders
are not included anywhere.  But first, we need to move everything out of
this header to avoid circular imports.

Ref T777