Refactor: libcrmcommon: Drop re-definition of LOG_TRACE
On the surface, it looks as if this changes public API. However, we
require libqb v1.0.1 or later. qb/qblog.h has defined the LOG_TRACE
constant since at least v0.5.0 (see commit 568befd).

So LOG_TRACE is always defined when we reach the dropped code.

Note that on GitHub, it looks as if commit 568befd was introduced in
v1.9.0. However, that's due to a tag naming mistake: the commit is
present in tag "1.9.0", which comes before "v1.9.0" in GitHub's sort
order. If you view all tags in which this commit is found, you'll see
that it's present in v0.5.0.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>