Refactor: various: Drop calls to CRM_TRACE_INIT_DATA()
With a clang compiler, this is a no-op.

With a non-clang compiler, this is defined as QB_LOG_INIT_DATA().
However, that macro did not actually initialize data. Rather, it is a
sanity check of some wildly complicated "linker magic" in libqb.

The libqb maintainers decided to drop the linker magic altogether in
v2.0.0, deciding that it provided little or no performance savings.
Beginning with v2.0.0, QB_LOG_INIT_DATA() is a no-op.

For a description of what QB_LOG_INIT_DATA() did and why it was dropped
(along with a LOT of related code), refer to the commit message and the
removed code/comments in commit 633f2624.
https://github.com/ClusterLabs/libqb/commit/633f2624

To summarize:
* Prior to libqb v2.0.0, QB_LOG_INIT_DATA() was only an optional sanity
  check (and was eventually documented accordingly).
* Starting with libqb v2.0.0, QB_LOG_INIT_DATA() did nothing at all.

So we lose nothing by dropping our calls to it.

Ref T1

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>