Refactor: libcrmcommon: Set blackbox file prefix in crm_log_init()
If this is a daemon. There's no need to wait to set it until we enable
(or disable, for some reason) the blackbox. We also make it non-volatile
(it can only be set once) and free it in crm_log_deinit().

Also cast getpid() to long long, since pid_t is a signed integer type.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>