Refactor: libcrmcommon: De-functionize logfile_disabled()
It's very short and only used once, and the name was misleading in my
opinion. It's more accurate to say that the argument is ignored. If
logging to some other file has been enabled, we don't disable it based
on these special values.

Also remove a redundant check against PCMK_VALUE_NONE in crm_log_init().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>