Doc: libcrmcommon: Clarify chown_logfile() comments
For whatever reason, the intent here is to ensure that the daemon group
owns the log file and can read and write it.

There are other ways that the daemon user could read and write the file.
For example, the file could belong to the daemon user even if it belongs
to a different group, or the group owner could be one of the daemon
user's secondary groups. It's not clear that we should be changing the
ownership in those cases. Also, the daemon group could have read/write
permissions via ACLs.

The current logic seems to try to ensure that ANY member of the daemon
group can read and write the file, although as discussed in the "\todo"
directive, it may not achieve that goal.

If we ever decide that we only want to ensure that the daemon user can
access the file, then we could impersonate the daemon user, try to open
the file, and then switch back to root. Or we could simply drop all of
this.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>