Fix: tools: cibadmin --xpath --delete-all deletes all matches again
This fixes a regression introduced by commit 11574193, which is not in
any release yet. The cibadmin_pre_delete_all function was incorrectly
attached to the delete command rather than to the delete_all command. As
a result, the cib_multiple flag was no longer getting set for
--delete-all.

Presumably this also made the behavior of cibadmin --xpath --delete
incorrect, deleting all matches instead of only the first one.

Fixes RHEL-131741

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>