Refactor: daemons: Simplify attrd message handlers a bit.
If the handler only deals with one type of message (IPC or CPG), it
should first check for the error case and handle that.  If it deals with
both types of messages, it should first check for whichever one requires
less code to process.  This allows as much of the body of the function
as possible to be indented as little as possible.