Refactor: daemons: Split out checking for a previous instance.
This does lose some of the existing error handling - if there's an OOM
condition, this code will no longer abort and it will instead be up to
whatever next allocates memory to fail.  However, this is almost exactly
what's happening in attrd and OOM should be extremely rare.

It's expected that eventually all daemons will use identical code for
this purpose, at which point we can handle these rare error cases
better.