Refactor: daemons: Un-indent the outermost block in get_action_timeout.
This is the first of many, many commits to do the trick where we invert
the condition, return early, and un-indent everything that comes after
it.  In this function there's actually a nested block that needs the
same trick applied to it, but that'll be a separate commit for ease of
review.