Refactor: fencer: Make spacing more consistent in single_test() macro
And drop "else" when crm_exit() makes it redundant.

The fact that (num_notifications - expected_notifications) is the number
of notifications that were received is not at all clear. But looking at
the way single_test() sets these values and st_callback() decrements
expected_notifications, it seems correct. There is room for improvement.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>