Refactor: libcrmcommon: pcmk__quote_cmdline() takes const char *const *
This is unfortunately the way to make a double-const (char **) (that is,
the top-level pointer and the next-level pointers are const). It also
unfortunately requires explicit casts. That was the initial motivation
for all of the output setup shuffling in previous commits. I wanted
fewer calls to pcmk__quote_cmdline().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>