Refactor: libcrmcommon: New assert_hr_format_alt()
This is a step toward converting the assertion helpers into macros. We'll
need multiple macros to avoid complaints from the compiler or Coverity
about NULL arguments, always-true/always-false conditions, etc.

In this case we could probably use variadic arguments, but the error
messages on failed assertions would be less clear. At best, we could
show that the result is not equal to the last expected string (the
alternate).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>