Refactor: libcrmcommon: Split part of assert_comment() into macro helper
This is a step toward replacing assert_comment() with a macro
completely. With a function, the CMocka assert macros show unhelpful
line numbers on error.

We'll need to split up assert_comment() into a NULL case and a non-NULL
case. The code that gets split out in this commit will be shared between
them.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>