Refactor: various: Use const char *const * where possible
In the remaining places I could find where it hasn't already been done.

This is annoying to look at, but it's correct. char ** and gchar ** are
obviously not const, and we should use const where we can. The
frustrating thing is that this pointer type is not compatible with
char **, so it requires explicit casts.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>