Refactor: based: Archive CIB file and digest file in subdirectory
Previously, in case of a digest mismatch, we renamed the CIB file and
the digest file in cib_root to new unique names generated by mkstemp().
Without checking the error logs, the file names did not make it clear
that the two files were related, or which one was the CIB file and which
one was the digest file.

Now, we create a new subdirectory of cib_root using mkdtemp(). Then we
move both the CIB file and the digest file to the new subdirectory. They
keep the base names "cib.xml" and "cib.xml.sig".

It's not necessary for archive_on_digest_mismatch() to take arguments,
but this avoids the need to allocate redundant strings for the old
paths.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>